# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Farming Service API",
        "description": "API documentation for Farming Service",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/service-areas": {
            "get": {
                "tags": [
                    "Service Areas"
                ],
                "summary": "Get all service areas",
                "operationId": "570649393246232ca3440cb3fbfb3d12",
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Service Areas"
                ],
                "summary": "Create a new service area",
                "operationId": "bd76742c6ab362e23384c4f96f8422a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id",
                                    "area_id"
                                ],
                                "properties": {
                                    "service_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "area_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/201"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/service-areas/{id}": {
            "get": {
                "tags": [
                    "Service Areas"
                ],
                "summary": "Get a specific service area",
                "operationId": "4ee6760ed407d65101e1514df752f02d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service Area ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Service Areas"
                ],
                "summary": "Update an existing service area",
                "operationId": "b8535a1987cf14b164174e13fb051842",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service Area ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "service_id",
                                    "area_id"
                                ],
                                "properties": {
                                    "service_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "area_id": {
                                        "type": "integer",
                                        "example": 4
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "422": {
                        "$ref": "#/components/responses/422"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Service Areas"
                ],
                "summary": "Soft delete a service area",
                "operationId": "0dea6db11623c89f50d282b4ea687577",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Service Area ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/200"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    },
                    "500": {
                        "$ref": "#/components/responses/500"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "responses": {
...
            "500": {
                "description": "Internal Server Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Something went wrong!"
                                },
                                "error": {
                                    "type": "string",
                                    "example": "Internal Server Error"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "401": {
                "description": "Invalid credentials",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Invalid credentials"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "200": {
                "description": "Successful operation",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Operation successful"
                                },
                                "data": {
                                    "type": "object",
                                    "example": []
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "201": {
                "description": "Resource created successfully",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Resource created successfully"
                                },
                                "data": {
                                    "type": "object",
                                    "example": []
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "422": {
                "description": "Validation error - invalid input data",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": false
                                },
                                "error": {
                                    "type": "string",
                                    "example": "The given data was invalid."
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "404": {
                "description": "Record not updated",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "status": {
                                    "type": "boolean",
                                    "example": true
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Not Updated!!"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}